1.1.43
In this version, full support for all keys has been implemented.
The handling of the Learn and Rate1 functions has been changed to the bottom keys [SNAP, TAP, OVERVIEW, SHIFT].

Encoders in CHANNEL mode smoothly control the speed of Master 1-4.

The keys [SNAP, TAP, OVERVIEW, SHIFT] control the Learn function for the Speed Master.

The buttons can blink at the BPM frequency of the Speed Master - in this case, you should assign Speed Master 1-4 to Executors 716-713. If you assign them correctly, the keys will light up yellow.
If the BPM key is lit yellow, it means the BPM is lower or higher than 60 - the key will blink. If the Speed Master is in the STOP position, the key will light up red.

If you press and hold the ACCENT key and then click [SNAP, TAP, OVERVIEW, or SHIFT], the keys will reset the BPM (Rate1) for the corresponding Speed Master.